move actionbar into courses and home fragment
