+++ app/src/main/kotlin/com/contentful/tea/kotlin/MainActivity.kt
@@ -3,11 +3,8 @@ package com.contentful.tea.kotlin
 import android.os.Bundle
 import android.view.Menu
 import android.view.MenuItem
-import androidx.annotation.IdRes
 import androidx.appcompat.app.AppCompatActivity
-import androidx.navigation.NavController
 import androidx.navigation.Navigation.findNavController
-import kotlinx.android.synthetic.main.activity_main.*
 
 class MainActivity : AppCompatActivity() {
 
@@ -18,10 +15,6 @@ class MainActivity : AppCompatActivity() {
 
         supportActionBar?.apply {
             title = ""
-        }
-
-        main_bottom_navigation.setOnNavigationItemSelectedListener {
-            bottomNavigationItemSelected(it)
         }
     }
 
@@ -43,26 +36,4 @@ class MainActivity : AppCompatActivity() {
 
     override fun onSupportNavigateUp() =
         findNavController(this, R.id.navigation_host_fragment).navigateUp()
-
-    private fun bottomNavigationItemSelected(item: MenuItem): Boolean {
-        val navController = findNavController(this, R.id.navigation_host_fragment)
-        return when (item.itemId) {
-            R.id.bottom_navigation_home -> {
-                navigateIfNotAlreadyThere(navController, R.id.home)
-            }
-            R.id.bottom_navigation_courses -> {
-                navigateIfNotAlreadyThere(navController, R.id.courses)
-                true
-            }
-            else -> false
-        }
-    }
-
-    private fun navigateIfNotAlreadyThere(navController: NavController, @IdRes id: Int): Boolean =
-        if (navController.currentDestination.id != id) {
-            navController.navigate(id)
-            true
-        } else {
-            false
-        }
 }
