+++ app/src/main/kotlin/com/contentful/tea/kotlin/courses/CoursesFragment.kt
@@ -2,11 +2,13 @@ package com.contentful.tea.kotlin.courses
 
 import android.os.Bundle
 import android.view.LayoutInflater
+import android.view.MenuItem
 import android.view.View
 import android.view.ViewGroup
+import androidx.annotation.IdRes
 import androidx.fragment.app.Fragment
+import androidx.navigation.NavController
 import androidx.navigation.Navigation
-import androidx.navigation.fragment.NavHostFragment
 import com.contentful.tea.kotlin.R
 import com.contentful.tea.kotlin.extensions.forEach
 import com.google.android.material.tabs.TabLayout
@@ -20,30 +22,39 @@ class CoursesFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        val navController = NavHostFragment.findNavController(this)
-        val root = inflater.inflate(R.layout.fragment_courses, container, false)
         arguments?.apply {
             categoryId = CoursesFragmentArgs.fromBundle(arguments).categoryId
         }
 
-        var navigation = root.findViewById<TabLayout>(R.id.top_navigation)
-        navigation.addOnTabSelectedListener(UsableOnTabListener { topNavigationItemSelected(it) })
-
-        // TODO: Make dynamic
-        val list: ViewGroup = root.findViewById(R.id.courses_list)
-        list.forEach { index, view ->
-            view.setOnClickListener {
-                val action = CoursesFragmentDirections.openCourseOverview("course " + index)
-                navController.navigate(action)
-            }
-        }
-        return root
+        return inflater.inflate(R.layout.fragment_courses, container, false)
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         // TODO: Make dynamic
-        top_navigation.addTab(top_navigation.newTab().setText(categoryId))
+        courses_top_navigation.addTab(courses_top_navigation.newTab().setText(categoryId))
+
+        courses_bottom_navigation.setOnNavigationItemSelectedListener {
+            if (activity != null) {
+                bottomNavigationItemSelected(it)
+                true
+            } else {
+                false
+            }
+        }
+        courses_top_navigation.addOnTabSelectedListener(UsableOnTabListener {
+            topNavigationItemSelected(
+                it
+            )
+        })
+
+        val navController = Navigation.findNavController(activity!!, R.id.navigation_host_fragment)
+        courses_list.forEach { index, view ->
+            view.setOnClickListener {
+                val action = CoursesFragmentDirections.openCourseOverview("course " + index)
+                navController.navigate(action)
+            }
+        }
     }
 
     private fun topNavigationItemSelected(tab: TabLayout.Tab) = if (!isDetached) {
@@ -51,6 +62,29 @@ class CoursesFragment : Fragment() {
         navController.navigate(CoursesFragmentDirections.openCategory(tab.tag as String))
     } else {
     }
+
+    private fun bottomNavigationItemSelected(item: MenuItem): Boolean {
+        val navController = Navigation.findNavController(activity!!, R.id.navigation_host_fragment)
+        return when (item.itemId) {
+            R.id.bottom_navigation_home -> {
+                navigateIfNotAlreadyThere(navController, R.id.home)
+                true
+            }
+            R.id.bottom_navigation_courses -> {
+                navigateIfNotAlreadyThere(navController, R.id.courses)
+                true
+            }
+            else -> false
+        }
+    }
+
+    private fun navigateIfNotAlreadyThere(navController: NavController, @IdRes id: Int): Boolean =
+        if (navController.currentDestination.id != id) {
+            navController.navigate(id)
+            true
+        } else {
+            false
+        }
 }
 
 class UsableOnTabListener(private val selectDelegate: (TabLayout.Tab) -> Unit) :
